refactor: implement new container API (#73)
This PR is objectively huge.

The main goal here was to refactor containers to be more aligned to what
server components bring to the table. Migrating to server components
proved to be a bridge to far mostly because MUI and RSC dont play good
enough yet, and the code does not benefit from RSC that much.

New containers are now forced to have 2 extra functions, `useData` and
`getData`. These replicate the functionality of `getData` of RSCs.


Completeley unrelated this PR refactors the core QB module to be more
generic.